# web/openapi.json
{
  "components": {
    "schemas": {
      "Activity": {
        "description": "The activity a track belongs to.\n\nActivity is orthogonal to :class:`~trackvault.domain.track_kind.TrackKind`: a\nplanned cycling route and a recorded cycling ride share this value and differ\nonly in kind.\n\nThe taxonomy is deliberately small and flat -- no sport hierarchy. Values come\nfrom explicit source metadata or from the user; the classifier does not guess\nthem, so ``UNKNOWN`` is a normal and permanently valid outcome. ``WALKING``\nand ``HIKING`` are separate values because the distinction matters for\nstatistics, but nothing may infer one from the other without evidence.",
        "enum": [
          "walking",
          "hiking",
          "cycling",
          "running",
          "scooter",
          "motorcycle",
          "other",
          "unknown"
        ],
        "title": "Activity",
        "type": "string"
      },
...
      "AggregationScope": {
        "description": "Which set of tracks a total covers.\n\nThere is deliberately no \"everything\" scope. A number that adds routes\nsomebody planned to distances somebody travelled is not a statistic about\neither, and making it the default would make it the one people quote.\n\nAttributes:\n    RECORDED: Tracks whose effective kind is ``RECORDED`` -- what actually\n        happened.\n    PLANNED: Tracks whose effective kind is ``PLANNED``.\n    UNKNOWN: Tracks whose kind could not be decided. Their own category, so\n        that a data-quality view is possible without them contaminating the\n        other two.",
        "enum": [
          "recorded",
          "planned",
          "unknown"
        ],
        "title": "AggregationScope",
        "type": "string"
      },
...
      "AvailableYearsResponse": {
        "description": "Which years this archive has something to show for.\n\nDeliberately not a range: an interface that offered \"this year and the\neleven before it\" opened an archive of 2019 recordings on an empty 2026,\nand it computed a calendar the archive already owns.",
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "archive_track_count": {
            "description": "Tracks held in any scope: what tells an empty archive from an empty selection",
            "title": "Archive Track Count",
            "type": "integer"
          },
          "scope": {
            "$ref": "#/components/schemas/AggregationScope"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "unplaced": {
            "$ref": "#/components/schemas/UnplacedCountsResponse",
            "description": "Tracks in scope that belong to no year -- reported beside it, never inside"
          },
          "years": {
            "description": "Local years the scope holds tracks in, newest first",
            "items": {
              "type": "integer"
            },
            "title": "Years",
            "type": "array"
          }
        },
        "required": [
          "scope",
          "activity",
          "timezone",
          "years",
          "unplaced",
          "archive_track_count"
        ],
        "title": "AvailableYearsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonthlyResponse": {
        "description": "A year broken into its twelve months.",
        "properties": {
          "activities": {
            "description": "The activities this year holds, in the taxonomy's order -- not the whole vocabulary, and never ordered by size. A client may rely on an activity keeping its place, and therefore its colour, across requests.",
            "items": {
              "$ref": "#/components/schemas/Activity"
            },
            "title": "Activities",
            "type": "array"
          },
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/MonthResponse"
            },
            "title": "Months",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/AggregationScope"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "scope",
          "activity",
          "timezone",
          "months",
          "activities"
        ],
        "title": "MonthlyResponse",
        "type": "object"
      },
      "OverallResponse": {
        "description": "Everything one scope holds, and each year inside it.\n\nA year is reported as twelve months because a calendar has twelve. This is\nreported as the years the archive has something for: padding out to the\nsupported calendar would draw a century of empty buckets to say nothing.",
        "properties": {
          "activities": {
            "description": "The activities this scope holds, in the taxonomy's order",
            "items": {
              "$ref": "#/components/schemas/Activity"
            },
            "title": "Activities",
            "type": "array"
          },
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope": {
            "$ref": "#/components/schemas/AggregationScope"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsResponse",
            "description": "What the whole scope adds up to. The years add up to exactly this."
          },
          "unplaced": {
            "$ref": "#/components/schemas/UnplacedResponse",
            "description": "Tracks in scope that belong to no year -- beside the total, never inside"
          },
          "years": {
            "description": "One bucket per year the archive holds something for, oldest first",
            "items": {
              "$ref": "#/components/schemas/YearBucketResponse"
            },
            "title": "Years",
            "type": "array"
          }
        },
        "required": [
          "scope",
          "activity",
          "timezone",
          "totals",
          "unplaced",
          "years",
          "activities"
        ],
        "title": "OverallResponse",
        "type": "object"
      },
...
      "UnplacedCountsResponse": {
        "description": "How many tracks belong to no calendar period, split by why.\n\nThe two halves are different facts and the second is the one that looks\nanswered: no instants at all, against instants nothing showed to be\nmeasured.",
        "properties": {
          "with_unverified_date": {
            "title": "With Unverified Date",
            "type": "integer"
          },
          "without_date": {
            "title": "Without Date",
            "type": "integer"
          }
        },
        "required": [
          "without_date",
          "with_unverified_date"
        ],
        "title": "UnplacedCountsResponse",
        "type": "object"
      },
      "UnplacedResponse": {
        "description": "What a scope holds that belongs to no calendar period.\n\nReported beside a year rather than inside it, and split by *why* a track has\nno period. A route with no instants and a route whose instants nothing\nvouches for are different facts about the archive, and only the second one\nlooks like a date until somebody checks.",
        "properties": {
          "with_unverified_date": {
            "$ref": "#/components/schemas/TotalsResponse",
            "description": "Tracks whose instants were not shown to have been measured"
          },
          "without_date": {
            "$ref": "#/components/schemas/TotalsResponse",
            "description": "Tracks carrying no instants at all"
          }
        },
        "required": [
          "without_date",
          "with_unverified_date"
        ],
        "title": "UnplacedResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YearBucketResponse": {
        "description": "One year of the whole archive.",
        "properties": {
          "by_activity": {
            "description": "The same partition a month reports, one level up",
            "items": {
              "$ref": "#/components/schemas/ActivityTotalsResponse"
            },
            "title": "By Activity",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsResponse"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "totals",
          "by_activity"
        ],
        "title": "YearBucketResponse",
        "type": "object"
      },
      "YearResponse": {
        "description": "One year's totals, in one scope.",
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope": {
            "$ref": "#/components/schemas/AggregationScope"
          },
          "timezone": {
            "description": "IANA zone the year's boundaries were drawn in",
            "title": "Timezone",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsResponse",
            "description": "Summed over the tracks whose activity date the archive can vouch for"
          },
          "unplaced": {
            "$ref": "#/components/schemas/UnplacedResponse",
            "description": "Tracks in scope that belong to no calendar period, and why"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "scope",
          "activity",
          "timezone",
          "totals",
          "unplaced"
        ],
        "title": "YearResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Self-hosted archive for recorded and planned activity tracks.",
    "title": "TrackVault",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/statistics/overall": {
      "get": {
        "description": "Return what a whole scope adds up to, and what each of its years did.\n\nRegistered before ``/year/{year}`` for the same reason ``/years`` is: a\nliteral path must win over a parameterised one whatever the router's\nmatching order turns out to be.",
        "operationId": "read_overall_api_v1_statistics_overall_get",
        "parameters": [
          {
            "description": "Which set of tracks to total: recorded, planned or unknown",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationScope",
              "default": "recorded",
              "description": "Which set of tracks to total: recorded, planned or unknown"
            }
          },
          {
            "description": "Narrow the totals to one activity",
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Activity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow the totals to one activity",
              "title": "Activity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Total every year at once",
        "tags": [
          "statistics"
        ]
      }
    },
    "/api/v1/statistics/year/{year}": {
      "get": {
        "description": "Return one year's totals.\n\nThe default scope is ``recorded``, because the unqualified question is what\nactually happened rather than what everything adds up to.",
        "operationId": "read_year_api_v1_statistics_year__year__get",
        "parameters": [
          {
            "description": "Year in the configured zone",
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "description": "Year in the configured zone",
              "maximum": 9999,
              "minimum": 1970,
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "description": "Which set of tracks to total: recorded, planned or unknown",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationScope",
              "default": "recorded",
              "description": "Which set of tracks to total: recorded, planned or unknown"
            }
          },
          {
            "description": "Narrow the totals to one activity",
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Activity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow the totals to one activity",
              "title": "Activity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YearResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Total one year",
        "tags": [
          "statistics"
        ]
      }
    },
    "/api/v1/statistics/year/{year}/monthly": {
      "get": {
        "description": "Return twelve monthly totals, including the months with no activity.",
        "operationId": "read_monthly_api_v1_statistics_year__year__monthly_get",
        "parameters": [
          {
            "description": "Year in the configured zone",
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "description": "Year in the configured zone",
              "maximum": 9999,
              "minimum": 1970,
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "description": "Which set of tracks to total: recorded, planned or unknown",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationScope",
              "default": "recorded",
              "description": "Which set of tracks to total: recorded, planned or unknown"
            }
          },
          {
            "description": "Narrow the totals to one activity",
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Activity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow the totals to one activity",
              "title": "Activity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Total each month of one year",
        "tags": [
          "statistics"
        ]
      }
    },
    "/api/v1/statistics/years": {
      "get": {
        "description": "Return the years one scope holds, newest first.\n\nRegistered before ``/year/{year}`` so that the literal path wins over the\nparameterised one whatever the router's matching order turns out to be.",
        "operationId": "read_available_years_api_v1_statistics_years_get",
        "parameters": [
          {
            "description": "Which set of tracks to total: recorded, planned or unknown",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationScope",
              "default": "recorded",
              "description": "Which set of tracks to total: recorded, planned or unknown"
            }
          },
          {
            "description": "Narrow the totals to one activity",
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Activity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow the totals to one activity",
              "title": "Activity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableYearsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the years this archive holds tracks in",
        "tags": [
          "statistics"
        ]
      }
    },
...
  }
}